refactor: add popups instead of 2 dialogs
