--- frontend/packages/app/src/pages/timesheet/components/submit-rejection/types.ts
@@ -1,4 +1,0 @@
-export type WeeklyRejectionProps = {
-  open: boolean;
-  onOpenChange: (open: boolean) => void;
-};
