+++ frontend/packages/app/src/pages/timesheet/team/weekly-approval/types.ts
@@ -1,6 +1,6 @@
 export type WeeklyApprovalProps = {
-  open: boolean;
-  onOpenChange: (open: boolean) => void;
   employee: string;
   startDate: string;
+  open: boolean;
+  onOpenChange: (open: boolean) => void;
 };
